refactor : basic cleanup of App.tsx
