+++ src/BotChat.ts
@@ -1,6 +1,7 @@
-export { App, AppProps, MakeId } from './App';
+export { App, AppProps} from './App';
 export { Chat, ChatProps } from './Chat';
 export * from 'botframework-directlinejs';
 export { queryParams } from './Attachment';
 export { SpeechOptions } from './SpeechOptions'
 export { Speech } from './SpeechModule'
+export { generateUserId } from './utils/generateUserId'
